Replace `jl_catch!` with `except!` and `rethrow!`
The old `jl_catch!` was replaced with two new macros for exception
handling. `except!` introduces three new block types.
`try { ... }` executes the code in the block. If any exceptions where
thrown in foreign Julia code, the thread `panic!`-s.
`catch <pattern> => { ... }` is used to catch the exceptions that would
otherwise cause a panic.
`finally { ... }` will execute arbitrary code after the executing the `try`
and `catch` blocks, even if one of the `catch` blocks returned from the
function.

`rethrow!(...)` can be used as a conveniency to return an
`Err(Error::UnhandledException(...))` in a function that returns a
`Result`.